Moved ArchiveBossContext to own folder
